Merge all IO functions in one (two)
Using a single function allows using the same code for the same device that is mapped in different IO spaces.
For instance, the NS16550 UART controller is port-mapped to 0x3F8 on IBM PC compatibles with uint8_t registers, while the same (compatible) controller on ARM or as a bus card may be memory-mapped with uint32_t registers. The code now can handle both at the same time using same driver code.